Last round of major changes.

+++ zazz/includes/standard/delete.php
@@ -14,23 +14,13 @@ if(defined('DEVELOPER')) {
 	}
 	
 	$pdo = Database::get()->PDO();
-	$q = $pdo->prepare('SHOW TABLES');
-	$q->execute();	
-	$rows = $q->fetchAll(PDO::FETCH_COLUMN, 0);
-	
-	foreach($rows as $row) {
-		$q = $pdo->prepare('DELETE FROM ' . $row);
-		$q->execute();			
-	}
-	
-	$pdo = Database::get()->PDO();
 	$q = $pdo->prepare('SHOW DATABASES');
 	$q->execute();	
 	$rows = $q->fetchAll(PDO::FETCH_COLUMN, 0);	
 	
 	foreach($rows as $row) {
 		if($row !== 'mysql' && $row !== 'performance_schema' && $row !== 'information_schema' 
-			&& $row !== 'test') {
+			&& $row !== 'test' && $row !== 'zazz') {
 			$q = $pdo->prepare('DROP DATABASE ' . $row);
 			$q->execute();
 		}
